refactor(progress): add src, tests, and types folders
