--- packages/progress/Progress.d.ts
@@ -1,14 +1,0 @@
-export interface ProgressProps {
-    tag?: React.ReactType | string;
-    animated?: boolean;
-    striped?: boolean;
-    complete?: boolean;
-    value?: number | string;
-    max?: number | string;
-    className?: string;
-    color?: string;
-}
-
-declare const Progress: React.FunctionComponent<ProgressProps>;
-
-export default Progress;
