+++ packages/progress/types/Progress.d.ts
@@ -1,0 +1,14 @@
+export interface ProgressProps {
+  tag?: React.ReactType | string;
+  animated?: boolean;
+  striped?: boolean;
+  complete?: boolean;
+  value?: number | string;
+  max?: number | string;
+  className?: string;
+  color?: string;
+}
+
+declare const Progress: React.FC<ProgressProps>;
+
+export default Progress;
